+++ spec/integration/braintree/credit_card_verification_spec.rb
@@ -67,6 +67,28 @@ describe Braintree::CreditCardVerification, "search" do
       expect(result.credit_card_verification.processor_response_type).to eq(Braintree::ProcessorResponseTypes::Approved)
     end
 
+    it "creates a new verification for Visa ANI when account information inquiry is passed" do
+      verification_params = {
+        :credit_card => {
+          :expiration_date => "05/2032",
+          :number => Braintree::Test::CreditCardNumbers::Visa,
+        },
+        :options => {
+          :account_information_inquiry => "send_data",
+        }
+      }
+
+      result = Braintree::CreditCardVerification.create(verification_params)
+
+      expect(result).to be_success
+      expect(result.credit_card_verification.status).to eq(Braintree::CreditCardVerification::Status::Verified)
+      expect(result.credit_card_verification.processor_response_code).to eq("1000")
+      expect(result.credit_card_verification.processor_response_text).to eq("Approved")
+      expect(result.credit_card_verification.ani_first_name_response_code).to eq("I")
+      expect(result.credit_card_verification.ani_last_name_response_code).to eq("I")
+      expect(result.credit_card_verification.processor_response_type).to eq(Braintree::ProcessorResponseTypes::Approved)
+    end
+
     it "creates a new verification from external vault param" do
       verification_params = {
         :credit_card => {